+++ regions/shapes/tests/test_ellipse.py
@@ -1,6 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import numpy as np
-from numpy.testing import assert_allclose
+from numpy.testing import assert_allclose, assert_equal
 import pytest
 
 import astropy.units as u
@@ -79,6 +79,74 @@ class TestEllipsePixelRegion(BaseTestPixelRegion):
         reg = self.reg.rotate(PixCoord(2, 3), 90 * u.deg)
         assert_allclose(reg.center.xy, (1, 4))
         assert_allclose(reg.angle.to_value("deg"), 95)
+
+    @pytest.mark.parametrize('sync', (False, True))
+    def test_as_mpl_selector(self, sync):
+
+        plt = pytest.importorskip('matplotlib.pyplot')
+
+        data = np.random.random((16, 16))
+        mask = np.zeros_like(data)
+
+        ax = plt.subplot(1, 1, 1)
+        ax.imshow(data)
+
+        def update_mask(reg):
+            mask[:] = reg.to_mask(mode='subpixels', subpixels=10).to_image(data.shape)
+
+        # For now this will only work with unrotated ellipses. Once this works
+        # with rotated ellipses, the following exception check can be removed
+        # as well as the ``angle=0 * u.deg`` in the call to copy() below.
+        with pytest.raises(NotImplementedError,
+                           match='Cannot create matplotlib selector for rotated ellipse.'):
+            self.reg.as_mpl_selector(ax)
+
+        region = self.reg.copy(angle=0 * u.deg)
+
+        selector = region.as_mpl_selector(ax, callback=update_mask, sync=sync)  # noqa
+
+        from matplotlib.backend_bases import MouseEvent, MouseButton
+
+        x, y = ax.transData.transform([[7.3, 4.4]])[0]
+        ax.figure.canvas.callbacks.process('button_press_event',
+                                           MouseEvent('button_press_event',
+                                                      ax.figure.canvas, x, y,
+                                                      button=MouseButton.LEFT))
+        x, y = ax.transData.transform([[9.3, 5.4]])[0]
+        ax.figure.canvas.callbacks.process('motion_notify_event',
+                                           MouseEvent('button_press_event',
+                                                      ax.figure.canvas, x, y,
+                                                      button=MouseButton.LEFT))
+        x, y = ax.transData.transform([[9.3, 5.4]])[0]
+        ax.figure.canvas.callbacks.process('button_release_event',
+                                           MouseEvent('button_press_event',
+                                                      ax.figure.canvas, x, y,
+                                                      button=MouseButton.LEFT))
+
+        ax.figure.canvas.draw()
+
+        if sync:
+
+            assert_allclose(region.center.x, 8.3)
+            assert_allclose(region.center.y, 4.9)
+            assert_allclose(region.width, 2)
+            assert_allclose(region.height, 1)
+            assert_quantity_allclose(region.angle, 0 * u.deg)
+
+            assert_equal(mask, region.to_mask(mode='subpixels', subpixels=10).to_image(data.shape))
+
+        else:
+
+            assert_allclose(region.center.x, 3)
+            assert_allclose(region.center.y, 4)
+            assert_allclose(region.width, 4)
+            assert_allclose(region.height, 3)
+            assert_quantity_allclose(region.angle, 0 * u.deg)
+
+            assert_equal(mask, 0)
+
+        with pytest.raises(Exception, match='Cannot attach more than one selector to a region.'):
+            region.as_mpl_selector(ax)
 
 
 class TestEllipseSkyRegion(BaseTestSkyRegion):